Move show and index jobs implementation into service layer.
